Define the IMemory interface.
We could directly use an Array<T> (being T Byte or Short) were necessary, but our goal is to be able to modify the get and set accessors, and that is not possible otherwise. This interface is intended to represent contiguous regions of memory: the memory itself, the stack, and the Vi (V0..VF) registers (which are not "contiguos memory", but have sequential identifiers.