Change the way the time is accounted
The Clock class responsibility (not IClock!) is only to account the
'total' time. A new class, Chronometer (implementing
IChronometer : IClock), is in charge of accounting the emulator's
running time, in order to be able to pause and resume the execution,
through its 'stop' and 'start' methods. This way, all of the
ClockDividers that use the same IClock (Chronometers) are kept in sync.