Made some Keyboard methods internal
The client code only has access to 'pressKey' and 'releaseKey' keyboard
methods through its new interface IKeyboard. The actual keyboard
implementation is moved to the 'core' package, as it has some
behavior crucial for instruction execution.

Maybe I could have just marked the methods as internal instead of
moving the class outside the io package, but I prefer to completely
decouple the keyboard that the Cpu uses from the one that the client
code manipulates using an interface.