Fix errors in the instruction family 8000h
The section "when" section for the instruction family 8000h had some errors:

First, the braces for the section "when", were'nt neccesary.
Second, it was necessary to make decode with mask 0xF00F in order to get each
instruction correctly.
on the other hand, it was necessary to handle an exception for instructions
that are'nt of CHIP-8's 8000h family.