Add the 'ICoreKeyboard' interface
In an analogous way to the 'IUserKeyboard', this interface only contains
methods useful to the core of the library, like the methods that
controls the latching to the next key release.

Also, added the soundTimer as a member of the keyboard, as it needs to
directly change its register value based on whether the keyboard is
waiting for a key to be released or not.