Fix drwVxVyNibble tests (VF and zero-size draws)
At the drwVxVyNibbleTallestSpriteTest:

Take care of VF. It's unnecessary on the aligned specific test
as it just checks the use of the position
(both aligned and not aligned are used in the other one).

At the drwVxVyNibbleIllegalMemoryTest:

Even if the size is 0, no one should use the draw instruction
on a invalid memory address.